Further updates and fixes.
This commit should fix the error with the round score being displayed when all four dice are the same. Also, I THINK I fixed the error where the Round Score value was not displayed at the beginning of each round. This should (In theory) fix that issue, however, I'm compiling this in Linux, not Windows, so results may vary.